dupdated owners section to define the ownerssection props and pending owner following hte tag details
