+++ datahub-web-react/src/app/tags/CreateNewTagModal/types.ts
@@ -1,5 +1,3 @@
-import { OwnerEntityType } from '../../../types.generated';
-
 export type CreateNewTagModalProps = {
     open: boolean;
     onClose: () => void;
@@ -16,31 +14,10 @@ export interface ModalButton {
     isLoading?: boolean;
 }
 
-// Interface for pending owner
-export interface PendingOwner {
-    ownerUrn: string;
-    ownerEntityType: OwnerEntityType;
-    ownershipTypeUrn: string;
-}
-
 // Common styled components
 export const FormSection = {
     marginBottom: '16px',
 };
-
-// Owners section props
-export interface OwnersSectionProps {
-    selectedOwnerUrns: string[];
-    setSelectedOwnerUrns: React.Dispatch<React.SetStateAction<string[]>>;
-    pendingOwners: PendingOwner[];
-    setPendingOwners: React.Dispatch<React.SetStateAction<PendingOwner[]>>;
-}
-
-// Entities section props
-export interface EntitiesSectionProps {
-    selectedEntityUrns: string[];
-    setSelectedEntityUrns: React.Dispatch<React.SetStateAction<string[]>>;
-}
 
 // Create tag result interface
 export interface CreateTagResult {
